#8 - Move OpenGL interop types to another folder and namespace
