+++ Testing/VelaptorTests/Content/AtlasLoaderTests.cs
@@ -11,7 +11,7 @@ namespace VelaptorTests.Content
     using Velaptor.Content;
     using Velaptor.Content.Exceptions;
     using Velaptor.Graphics;
-    using Velaptor.NativeInterop;
+    using Velaptor.NativeInterop.OpenGL;
     using Velaptor.Services;
     using Xunit;
     using Assert = Helpers.AssertExtensions;
