+++ Testing/VelaptorTests/OpenGL/ShaderProgramTests.cs
@@ -6,7 +6,7 @@ namespace VelaptorTests.OpenGL
 {
     using System;
     using Moq;
-    using Velaptor.NativeInterop;
+    using Velaptor.NativeInterop.OpenGL;
     using Velaptor.OpenGL;
     using Velaptor.Services;
     using Xunit;
