+++ Velaptor/Content/FontLoader.cs
@@ -13,6 +13,7 @@ namespace Velaptor.Content
     using Velaptor.Content.Exceptions;
     using Velaptor.Graphics;
     using Velaptor.NativeInterop;
+    using Velaptor.NativeInterop.OpenGL;
     using Velaptor.Services;
 
     /// <summary>
