+++ Velaptor/OpenGL/GLWindowFacade.cs
@@ -6,10 +6,7 @@ namespace Velaptor.OpenGL
 {
 #pragma warning disable SA1135 // Using directives should be qualified
     using System;
-    using System.Collections.Generic;
-    using System.Diagnostics;
     using System.Diagnostics.CodeAnalysis;
-    using System.Linq;
     using System.Numerics;
     using Silk.NET.Input;
     using Silk.NET.Maths;
